+++ backend/canisters/multi_user/api/src/updates/mod.rs
@@ -1,0 +1,1 @@
+pub mod c2c_create_user;